Extract enemy and boss data loaders
